feat: sales services

+++ services/sale.service.js
@@ -7,10 +7,17 @@ const createSale = async (sale) => {
   const product = await ProductRepository.getProduct(sale.product_id);
   if (!client) throw new Error("Client not found");
   if (!product) throw new Error("Product not found");
-  return await SaleRepository.insertSale(sale);
+
+  if (product.stock > 0) {
+    sale = await SaleRepository.insertSale(sale);
+    product.stock--;
+    await ProductRepository.updateProduct(product);
+    return sale;
+  } else throw new Error("Product out of stock");
 };
 
-const getSales = async () => {
+const getSales = async (productId) => {
+  if (productId) return await SaleRepository.getSalesByProductId(productId);
   return await SaleRepository.getSales();
 };
 
@@ -19,6 +26,13 @@ const getSale = async (id) => {
 };
 
 const deleteSale = async (id) => {
+  const sale = await SaleRepository.getSale(id);
+  if (!sale) throw new Error("Sale not found");
+
+  const product = await ProductRepository.getProduct(sale.product_id);
+  product.stock++;
+  await ProductRepository.updateProduct(product);
+
   return await SaleRepository.deleteSale(id);
 };
